# api/v1/openapi.yaml
openapi: 3.0.0
info:
  title: "Specification for auth service"
  version: 1.0.0

package: api
generate:
  std-http-server: true
  models: true
output: gen.go

paths:
  /api/v1/auth/register:
    post:
      summary: "Method to register a new user."
      description: "Creates a new user by email and password"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: "Success response with data of created user"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: "Email already exists"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: "Unprocessable Entity"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      summary: "Method to login"
      description: "Creates a new tokens for user to access secure endpoints"
      requestBody:
        required: true
        content:
          application/json:
              schema:
                $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: "Successful login"
          headers:
            Set-Cookie:
              description: "Refresh token cookie"
              schema:
                type: string
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: "Unprocessable Entity"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/google/login:
    get:
      summary: "Start Google OAuth2 login (redirect to Google)"
      operationId: AuthGoogleLogin
      tags:
        - oauth
      responses:
        '302':
          description: Redirect to Google Authorization URL
          headers:
            Set-Cookie:
              description: "Set state cookie"
              schema:
                type: string
            Location:
              description: "URL to redirect user agent to (Google)"
              schema:
                type: string
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/google/callback:
    get:
      summary: "OAuth2 callback from Google"
      operationId: AuthGoogleCallback
      tags:
        - oauth
      parameters:
        - name: code
          in: query
          schema:
            type: string
          required: true
        - name: state
          in: query
          schema:
            type: string
          required: true
      responses:
        '200':
          description: "Successful authentication"
          headers:
            Set-Cookie:
              description: "Refresh token cookie"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '400':
          description: "Bad Request (invalid state / missing parameters)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/me:
    get:
      summary: "Secured method to get information about user"
      description: "Allows to obtaining information about an authorized user by access token"
      security:
        - BearerAuth: []
      responses:
        '200':
          description: "User information successfully retrieved"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      summary: "Secured method to logout"
      description: "Allows to logout user by refresh token"
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '204':
          description: "Successful user logout"
          headers:
            Set-Cookie:
              schema:
                type: string
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      summary: "Method to refresh access and refresh tokens"
      description: "Invalidates the previous refresh token and generates new access + refersh token"
      parameters:
        - name: refresh_token
          in: cookie
          required: true
          schema:
            type: string
      responses:
        '200':
          description: "Successful refresh access token"
          headers:
            Set-Cookie:
              description: "Refresh token cookie"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '400':
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '504':
          description: "Gateway Timeout"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.org"
        password:
          type: string
          minLength: 8
          maxLength: 100
          example: "password"
      required:
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@example.org"
        password:
          type: string
          example: "password"
      required:
        - email
        - password
    
    AccessToken:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
      required:
        - access_token
    
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426655440000"
        email:
          type: string
          example: "user@example.org"
        created_at:
          type: string
          format: date-time
          example: "2025-11-25T12:34:56Z"
      required:
        - user_id
        - email
        - created_at
    UserInfoResponse:
      type: object
      properties:
        user_id:
          type: string
          example: "123e4567-e89b-12d3-a456-426655440000"
        email:
          type: string
          example: "user@example.org"
        created_at:
          type: string
          format: date-time
          example: "2025-11-25T12:34:56Z"
      required:
        - user_id
        - email
        - created_at
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
      required:
        - status
        - message